refactor(domain): move FilePath and LanguageCode to shared kernel
Generic value objects that are used across bounded contexts now live
in domain/shared/value_objects/. Original locations re-export for
backward compatibility.
