+++ src/domain/shared/value_objects/__init__.py
@@ -1,0 +1,9 @@
+"""Shared value objects used across bounded contexts."""
+
+from src.domain.shared.value_objects.file_path import FilePath
+from src.domain.shared.value_objects.language_code import LanguageCode
+
+__all__ = [
+    "FilePath",
+    "LanguageCode",
+]
